Before: `BenchmarkAddPiece512M-32    	       1	7034478758 ns/op	  75.72 MB/s`

After: `BenchmarkAddPiece512M-32    	       2	 916931366 ns/op	 580.93 MB/s`
(on AMD TR2950X)