@Kubuxu 2021-01-11
Are you 100% that pieces can be combined and the resulting CID will be the same? Do we have tests for this?