@magik6k 2021-01-11
> Are you 100% that pieces can be combined and the resulting CID will be the same? Do we have tests for this?

Yep, It's a simple binary tree under the hood. (The CID in TestAddPiece512M was generated in a single call. Also, we are already doing this chunking on master, this is just making things run in parallel)