@magik6k 2021-01-12
(tested manually on my miner, works, and it's really fast there too)