@shaodan 2021-01-12
@magik6k Should we change resource that addpiece needs in [sector-storage/resources.go](https://github.com/filecoin-project/lotus/blob/master/extern/sector-storage/resources.go#L53)